+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/licence-application/components/business/steps-business-licence-information.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, ViewChild, ViewEncapsulation } from '@angular/core';
+import { Component, EventEmitter, Input, Output, ViewChild, ViewEncapsulation } from '@angular/core';
 import { Router } from '@angular/router';
 import { ApplicationTypeCode } from '@app/api/models';
 import { BaseWizardStepComponent } from '@app/core/components/base-wizard-step.component';
@@ -42,10 +42,27 @@ import { StepBusinessLicenceLiabilityComponent } from './step-business-licence-l
 				></app-wizard-footer>
 			</mat-step>
 
+			<mat-step *ngIf="isRenewal">
+				<app-step-business-licence-is-change-summary></app-step-business-licence-is-change-summary>
+
+				<div class="row wizard-button-row">
+					<div class="offset-xxl-4 col-xxl-5 offset-xl-3 col-xl-7 offset-lg-3 col-lg-7 col-md-12">
+						<button mat-flat-button color="primary" class="large mb-2" (click)="onRenewalLongFormNextStep()">
+							Yes, I need to update
+						</button>
+					</div>
+				</div>
+
+				<app-wizard-footer
+					nextButtonLabel="No updates needed"
+					[isWideNext]="true"
+					(previousStepperStep)="onGoToPreviousStep()"
+					(nextStepperStep)="onRenewalShortFormNextStep()"
+				></app-wizard-footer>
+			</mat-step>
+
 			<mat-step *ngIf="isNew">
-				<app-step-business-licence-expired
-					(validExpiredLicenceData)="onValidExpiredLicence()"
-				></app-step-business-licence-expired>
+				<app-step-business-licence-expired></app-step-business-licence-expired>
 
 				<app-wizard-footer
 					(previousStepperStep)="onGoToPreviousStep()"
@@ -53,7 +70,7 @@ import { StepBusinessLicenceLiabilityComponent } from './step-business-licence-l
 				></app-wizard-footer>
 			</mat-step>
 
-			<mat-step>
+			<mat-step *ngIf="!isRenewalShortForm">
 				<app-step-business-licence-company-branding
 					[applicationTypeCode]="applicationTypeCode"
 				></app-step-business-licence-company-branding>
@@ -96,7 +113,10 @@ export class StepsBusinessLicenceInformationComponent extends BaseWizardStepComp
 	@Input() isBusinessLicenceSoleProprietor!: boolean;
 	@Input() isFormValid!: boolean;
 	@Input() showSaveAndExit!: boolean;
+	@Input() isRenewalShortForm!: boolean;
 	@Input() applicationTypeCode!: ApplicationTypeCode;
+
+	@Output() renewalShortForm: EventEmitter<boolean> = new EventEmitter();
 
 	@ViewChild(StepBusinessLicenceExpiredComponent) stepExpiredComponent!: StepBusinessLicenceExpiredComponent;
 	@ViewChild(StepBusinessLicenceCompanyBrandingComponent)
@@ -114,8 +134,12 @@ export class StepsBusinessLicenceInformationComponent extends BaseWizardStepComp
 		);
 	}
 
-	onValidExpiredLicence(): void {
-		this.childNextStep.emit(true);
+	onRenewalShortFormNextStep(): void {
+		this.renewalShortForm.emit(true);
+	}
+
+	onRenewalLongFormNextStep(): void {
+		this.renewalShortForm.emit(false);
 	}
 
 	override dirtyForm(step: number): boolean {
