+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/licence-application/components/business/steps-business-licence-review.component.ts
@@ -9,7 +9,7 @@ import { StepBusinessLicenceSummaryComponent } from './step-business-licence-sum
 	selector: 'app-steps-business-licence-review',
 	template: `
 		<mat-stepper class="child-stepper" (selectionChange)="onStepSelectionChange($event)" #childstepper>
-			<mat-step>
+			<mat-step *ngIf="!isRenewalShortForm">
 				<app-step-business-licence-summary (editStep)="onGoToStep($event)"></app-step-business-licence-summary>
 
 				<app-wizard-footer
@@ -31,7 +31,7 @@ import { StepBusinessLicenceSummaryComponent } from './step-business-licence-sum
 					[showSaveAndExit]="true"
 					(saveAndExit)="onNoSaveAndExit()"
 					[nextButtonLabel]="submitPayLabel"
-					(previousStepperStep)="onGoToPreviousStep()"
+					(previousStepperStep)="onConsentGoToPreviousStep()"
 					(nextStepperStep)="onPayNow()"
 				></app-wizard-footer>
 			</mat-step>
@@ -45,6 +45,7 @@ export class StepsBusinessLicenceReviewComponent extends BaseWizardStepComponent
 
 	@Input() workerLicenceTypeCode!: WorkerLicenceTypeCode;
 	@Input() applicationTypeCode!: ApplicationTypeCode;
+	@Input() isRenewalShortForm!: boolean;
 
 	@Output() goToStep: EventEmitter<number> = new EventEmitter<number>();
 
@@ -61,6 +62,15 @@ export class StepsBusinessLicenceReviewComponent extends BaseWizardStepComponent
 		// if (this.applicationTypeCode === ApplicationTypeCode.Update) {
 		// this.submitPayLabel = 'Submit';
 		// }
+	}
+
+	onConsentGoToPreviousStep(): void {
+		if (this.isRenewalShortForm) {
+			this.onStepPrevious();
+			return;
+		}
+
+		this.onGoToPreviousStep();
 	}
 
 	onPayNow(): void {
