+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/licence-application/licence-application.module.ts
@@ -94,6 +94,7 @@ import { BusinessProfileComponent } from './components/business/business-profile
 import { BusinessUserApplicationsComponent } from './components/business/business-user-applications.component';
 import { CommonBusinessBcBranchesComponent } from './components/business/common-business-bc-branches.component';
 import { CommonBusinessInformationComponent } from './components/business/common-business-information.component';
+import { CommonBusinessLicenceSummaryComponent } from './components/business/common-business-licence-summary.component';
 import { CommonBusinessProfileComponent } from './components/business/common-business-profile.component';
 import { CommonControllingMembersComponent } from './components/business/common-controlling-members.component';
 import { CommonEmployeesComponent } from './components/business/common-employees.component';
@@ -113,6 +114,7 @@ import { StepBusinessLicenceControllingMemberInvitesComponent } from './componen
 import { StepBusinessLicenceControllingMembersComponent } from './components/business/step-business-licence-controlling-members.component';
 import { StepBusinessLicenceEmployeesComponent } from './components/business/step-business-licence-employees.component';
 import { StepBusinessLicenceExpiredComponent } from './components/business/step-business-licence-expired.component';
+import { StepBusinessLicenceIsChangeSummaryComponent } from './components/business/step-business-licence-is-change-summary.component';
 import { StepBusinessLicenceLiabilityComponent } from './components/business/step-business-licence-liability.component';
 import { StepBusinessLicenceManagerInformationComponent } from './components/business/step-business-licence-manager-information.component';
 import { StepBusinessLicenceProfileComponent } from './components/business/step-business-licence-profile.component';
@@ -232,6 +234,7 @@ import { LicenceApplicationService } from './services/licence-application.servic
 		CommonBusinessAddressComponent,
 		CommonBusinessBcBranchesComponent,
 		CommonBusinessInformationComponent,
+		CommonBusinessLicenceSummaryComponent,
 		CommonBusinessProfileComponent,
 		CommonBusinessTermsComponent,
 		CommonContactInformationComponent,
@@ -310,6 +313,7 @@ import { LicenceApplicationService } from './services/licence-application.servic
 		StepBusinessLicenceProfileComponent,
 		StepsBusinessLicenceReviewComponent,
 		StepBusinessLicenceSummaryComponent,
+		StepBusinessLicenceIsChangeSummaryComponent,
 		StepBusinessLicenceTermComponent,
 		StepBusinessLicenceExpiredComponent,
 		StepBusinessLicenceManagerInformationComponent,
