+++ src/Spd.Presentation.Licensing/ClientApp/src/app/modules/licence-application/services/business-application.service.ts
@@ -70,6 +70,7 @@ export class BusinessApplicationService extends BusinessApplicationHelper {
 
 		isBcBusinessAddress: new FormControl(), // placeholder for flag
 		isBusinessLicenceSoleProprietor: new FormControl(), // placeholder for flag
+		isRenewalShortForm: new FormControl(), // placeholder for flag
 
 		originalLicenceData: this.originalBusinessLicenceFormGroup,
